+++ lib/src/model/internal/internal_declaration_utils.dart
@@ -46,6 +46,10 @@ abstract class InternalDeclarationUtils {
     return result;
   }
 
+  static bool hasVisibleForTesting(Element element) {
+    return containsAnnotation(element, 'visibleForTesting');
+  }
+
   static bool hasExperimental(Element element) {
     return containsAnnotation(element, 'experimental');
   }
@@ -60,6 +64,39 @@ abstract class InternalDeclarationUtils {
       return path.relative(name, from: rootPath);
     }
     return '';
+  }
+
+  static String getFullQualifiedNameFor(Element element) {
+    final parts = <String>[];
+
+    /// stop at compilation unit level + adapt display name to show the relative path
+    if (element is CompilationUnitElement) {
+      Uri uri = element.source.uri;
+      if (uri.isScheme('file')) {
+        final pathParts = path.split(uri.toFilePath());
+        final libIndex = pathParts.lastIndexOf('lib');
+        if (libIndex >= 0) {
+          pathParts.removeRange(0, libIndex + 1);
+        }
+        return path.joinAll(pathParts);
+      }
+      // in case we have a URI we just show the path segments. If we happen to find "src" then we only show the path after that
+      // this works as we only have usages in our own package
+      final uriPathSegments = [...uri.pathSegments];
+      final srcIndex = uriPathSegments.lastIndexOf('src');
+      if (srcIndex >= 0) {
+        uriPathSegments.removeRange(0, srcIndex + 1);
+      }
+
+      return uriPathSegments.join('/');
+    }
+
+    if (element.enclosingElement != null) {
+      parts.add(getFullQualifiedNameFor(element.enclosingElement!));
+    }
+    parts.add(element.displayName);
+
+    return parts.join('::');
   }
 
   static String? getNamespaceForElement(