+++ lib/src/model/model.dart
@@ -10,3 +10,4 @@ export 'sdk_type.dart';
 export 'type_alias_declaration.dart';
 export 'type_hierarchy.dart';
 export 'type_usage.dart';
+export 'type_usage_kind.dart';