+++ test/test_packages/missing_export/package_a/lib/types/class_a.dart
@@ -1,5 +1,4 @@
 import 'package:meta/meta.dart';
-import 'package:package_b/package_b.dart';
 
 import 'class_b.dart';