+++ test/test_packages/missing_export_for_test/package_a/lib/types/class_a.dart
@@ -1,0 +1,11 @@
+import 'package:meta/meta.dart';
+
+import 'class_b.dart';
+
+@experimental
+class ClassA {
+  ClassA();
+
+  @visibleForTesting
+  ClassB createClassB() => ClassB('someValue');
+}